Feat(view): Add OutputView class and move InputView to view package
- Created OutputView with printLottoAmount method to display purchased lotto count
